requirements.txt
pandas>=1.3.0
numpy>=1.21.0
sqlalchemy>=1.4.0
python-dotenv>=0.19.0
openpyxl>=3.0.9
configparser>=5.0.2
python-dateutil>=2.8.2
pytz>=2021.3
requests>=2.26.0
pyodbc>=4.0.30
openai>=1.0.0
google-cloud-bigquery
sqlglot>=3.0.0
google-auth>=2.0.0
SQLGlot>=10.0.0
